+++ src/mkvtoolnix-gui/jobs/mux_job.cpp
@@ -198,7 +198,8 @@ MuxJob::config()
 }
 
 void
-MuxJob::runProgramSetupVariables(ProgramRunner::VariableMap &variables) {
+MuxJob::runProgramSetupVariables(ProgramRunner::VariableMap &variables)
+  const{
   Job::runProgramSetupVariables(variables);
 
   // OUTPUT_… are kept for backwards compatibility.